Migrate zod schemas to internal package
