+++ backend/package.json
@@ -24,7 +24,8 @@
     "lodash": "^4.17.21",
     "swagger-jsdoc": "^6.2.5",
     "swagger-ui-express": "^4.5.0",
-    "zod": "^3.19.1"
+    "zod": "^3.19.1",
+    "schemas": "*"
   },
   "devDependencies": {
     "@types/cors": "^2.8.12",
