+++ backend/src/middleware/validateRequest.ts
@@ -1,4 +1,4 @@
-import { Request, Response, NextFunction } from "express";
+import { NextFunction, Request, Response } from "express";
 import { AnyZodObject } from "zod";
 
 const validate =
