+++ backend/src/services/session.service.ts
@@ -1,3 +1,3 @@
 import prisma from "../utils/prisma.util";
 
-export async function createSession() {}
+// export async function createSession() {}
