+++ backend/src/swagger/session.swagger.ts
@@ -1,5 +1,3 @@
-import { object, string } from "zod";
-
 /**
  * @openapi
  * components:
@@ -24,14 +22,3 @@ import { object, string } from "zod";
  *      accessToken:
  *       type: string
  */
-
-export const createSessionSchema = object({
-  body: object({
-    email: string({
-      required_error: "Email is required",
-    }).email("Not a valid email"),
-    password: string({
-      required_error: "Password is required",
-    }),
-  }),
-});
