+++ packages/schemas/package.json
@@ -1,0 +1,23 @@
+{
+  "name": "schemas",
+  "version": "1.0.0",
+  "description": "",
+  "files": [
+    "dist/*"
+  ],
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "tsc",
+    "dev": "tsc -w --preserveWatchOutput"
+  },
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "zod": "^3.19.1"
+  },
+  "devDependencies": {
+    "typescript": "^4.8.4"
+  }
+}
