+++ packages/schemas/src/index.ts
@@ -1,0 +1,3 @@
+export * from "./movies.schema";
+export * from "./session.schema";
+export * from "./user.schema";
