+++ packages/schemas/src/movies.schema.ts
@@ -1,0 +1,9 @@
+import z from "zod";
+
+export const createMovieSchema = z.object({
+  body: z.object({
+    title: z.string({
+      required_error: "Title is required",
+    }),
+  }),
+});
