+++ packages/schemas/src/session.schema.ts
@@ -1,0 +1,12 @@
+import { object, string } from "zod";
+
+export const createSessionSchema = object({
+  body: object({
+    email: string({
+      required_error: "Email is required",
+    }).email("Not a valid email"),
+    password: string({
+      required_error: "Password is required",
+    }),
+  }),
+});
