+++ packages/schemas/src/user.schema.ts
@@ -1,0 +1,21 @@
+import { object, string } from "zod";
+
+export const createUserSchema = object({
+  body: object({
+    username: string({
+      required_error: "Username is required",
+    }),
+    password: string({
+      required_error: "Password is required",
+    }).min(8, "Password too short - should be 8 chars minimum"),
+    passwordConfirmation: string({
+      required_error: "passwordConfirmation is required",
+    }),
+    email: string({
+      required_error: "Email is required",
+    }).email("Not a valid email"),
+  }).refine((data) => data.password === data.passwordConfirmation, {
+    message: "Passwords do not match",
+    path: ["passwordConfirmation"],
+  }),
+});
